// src/app/core/models/userlevelcontrol.ts
export class UserlevelControl {
    SysID?: string;
    Password?: string;
    UserLevel?: string;
    Description?: string;
    RestrictedIplist?: string;
    Remark?: string;
    MenuList?: boolean;
    AccessLogin?: boolean;
    AccessMobileLogin?: boolean;
    IsAdministrator?: boolean;
    Dval?: string;
    MenuArrays?: number[];
}